CI: don't run sanity checks when tags are pushed
We already weren't running sanity checks on new tags in mesonbuild/wrapdb
because the tags are created automatically by a workflow, so GitHub's
recursion avoidance doesn't start the checks.  However, when manually
testing changes in a fork, it's necessary to keep the fork's tags
up-to-date to prevent sanity checks from testing all untagged wrap
updates, and pushing tags to the fork would start CI on each of them.
Avoid this by running sanity checks only on pull requests and branch
pushes.