sanity_checks: in PR CI, avoid tag check racing with tag creation
If a PR is merged and the create_release job creates a new tag, any
subsequent CI jobs on other PRs will fail, since the tag won't be
mentioned in those PRs' releases.json.  We do want the releases.json check
to be pretty broad, e.g. to catch removals of existing release versions.
However, it's annoying to have to rebase PRs solely because an unrelated
project was updated, especially since we land PRs using the "rebase and
merge" strategy anyway.

actions/checkout checks out a synthetic merge commit when running on a PR
(but not when running directly on a branch), whose timestamp appears to
match the last branch update.  If this commit exists, skip the
releases.json check for any tags whose timestamps are later than the one
from the merge commit.

(Actually our tags don't have their own timestamps, since we're using
lightweight tags, so check the committer timestamp of the underlying
commit.)